+++ src/app/(website)/courses/[...slug]/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata, ResolvingMetadata } from 'next'
 
-import { getCourseBySlug } from '@/services/courses/course-api'
+import { getCourseBySlug } from '@/services/course/course-api'
 
 import CourseDetailView from '@/sections/courses/view/course-detail-view'
